--- a/toolz/dicttoolz.py
+++ b/toolz/dicttoolz.py
@@ -197,7 +197,7 @@ def assoc(d, key, value, factory=dict):
     return d2
 
 
-def dissoc(d, *keys):
+def dissoc(d, *keys, **kwargs):
     """ Return a new dict with the given key(s) removed.
 
     New dict has d[key] deleted for each supplied key.
@@ -210,11 +210,20 @@ def dissoc(d, *keys):
     >>> dissoc({'x': 1}, 'y') # Ignores missing keys
     {'x': 1}
     """
-    d2 = copy.copy(d)
-    for key in keys:
-        if key in d2:
-            del d2[key]
-    return d2
+    factory = _get_factory(dissoc, kwargs)
+    d2 = factory()
+
+    if len(keys) < len(d) * .6:
+        d2.update(d)
+        for key in keys:
+            if key in d2:
+                del d2[key]
+    else:
+        remaining = set(d)
+        remaining.difference_update(keys)
+        for k in remaining:
+            d2[k] = d[k]
+        return d2
 
 
 def assoc_in(d, keys, value, factory=dict):