--- a/toolz/tests/test_dicttoolz.py
+++ b/toolz/tests/test_dicttoolz.py
@@ -90,16 +90,16 @@ class TestDict(object):
 
     def test_dissoc(self):
         D, kw = self.D, self.kw
-        assert dissoc(D({"a": 1}), "a") == D({})
-        assert dissoc(D({"a": 1, "b": 2}), "a") == D({"b": 2})
-        assert dissoc(D({"a": 1, "b": 2}), "b") == D({"a": 1})
-        assert dissoc(D({"a": 1, "b": 2}), "a", "b") == D({})
-        assert dissoc(D({"a": 1}), "a") == dissoc(dissoc(D({"a": 1}), "a"), "a")
+        assert dissoc(D({"a": 1}), "a", **kw) == D({})
+        assert dissoc(D({"a": 1, "b": 2}), "a", **kw) == D({"b": 2})
+        assert dissoc(D({"a": 1, "b": 2}), "b", **kw) == D({"a": 1})
+        assert dissoc(D({"a": 1, "b": 2}), "a", "b", **kw) == D({})
+        assert dissoc(D({"a": 1}), "a", **kw) == dissoc(dissoc(D({"a": 1}), "a", **kw), "a", **kw)
 
         # Verify immutability:
         d = D({'x': 1})
         oldd = d
-        d2 = dissoc(d, 'x')
+        d2 = dissoc(d, 'x', **kw)
         assert d is oldd
         assert d2 is not oldd