Dissoc was really fast when the keys to remove were relatively few
compared to the size of the dictionary.  This version will pick the smaller
container to iterate over, either building up a new dictionary or tearing
down a copy.

A small heuristic is used to decide if the overhead of building a set is
worth the effort.  My rudimentry testing shows that if the number of keys is
less than 60% the length of the dictionary, tearing down a copy is faster.
If keys is very large, we calculate the difference between mapping and keys 
and build a new dictionary out of that difference.

This PR trys to change the signature of `dissoc` in a backwards compatible way by adding the `factory` keyword argument.  This makes the function consistent with the call signatures of all the other functions that mutate mappings.  This change was required  otherwise, we'd be unable to initialize a new empty mapping from a factory.  If this is undesirable, we can resort to using `d.__class__()` as the factory.

Microbenchmarks coming soon...

Do you have any feedback, @eriknw?