@groutr 2018-11-16
@eriknw, thanks.  I'll update the tests.

Computing the remaining keys benefits a lot from being able to use mapping views (which are set-like).  In Python 3, custom mapping classes automatically get these set-like views.  Does toolz anticipate dropping Python 2 support in 2020?