@groutr 2018-12-06
ping @eriknw.  This is ready for review.